Refactor: Remove unused files and merge planning solutions computation methods
