+++ solution/planning_solutions.go
@@ -19,6 +19,49 @@ const (
 	CategorizedPlaceIterInitFailureErrMsg = "categorized places iterator init failure"
 )
 
+type PlanningSolution struct {
+	PlaceNames      []string            `json:"place_names"`
+	PlaceIDS        []string            `json:"place_ids"`
+	PlaceLocations  [][2]float64        `json:"place_locations"` // lat,lng
+	PlaceAddresses  []string            `json:"place_addresses"`
+	PlaceURLs       []string            `json:"place_urls"`
+	PlaceCategories []POI.PlaceCategory `json:"place_categories"`
+	Score           float64             `json:"score"`
+}
+
+func CreatePlanningSolutionCandidate(iter MultiDimIterator, placeClusters [][]matching.Place) (res PlanningSolution) {
+	if len(iter.Status) != len(placeClusters) {
+		return
+	}
+	// deduplication of repeating places in the result
+	record := make(map[string]bool)
+	places := make([]matching.Place, len(iter.Status))
+	for idx, placeIdx := range iter.Status {
+		placesByCategory := placeClusters[idx]
+
+		var place = placesByCategory[placeIdx]
+
+		// if the same place appears in two indexes, return incomplete result
+		if _, exist := record[place.GetPlaceId()]; exist {
+			return
+		}
+
+		record[place.GetPlaceId()] = true
+		places[idx] = place
+		res.PlaceIDS = append(res.PlaceIDS, place.GetPlaceId())
+		res.PlaceNames = append(res.PlaceNames, place.GetPlaceName())
+		res.PlaceLocations = append(res.PlaceLocations, [2]float64{place.GetLocation().Latitude, place.GetLocation().Longitude})
+		res.PlaceAddresses = append(res.PlaceAddresses, place.GetPlaceFormattedAddress())
+		res.PlaceCategories = append(res.PlaceCategories, place.GetPlaceCategory())
+		if len(strings.TrimSpace(place.GetURL())) == 0 {
+			place.SetURL(iowrappers.GoogleSearchHomePageURL)
+		}
+		res.PlaceURLs = append(res.PlaceURLs, place.GetURL())
+	}
+	res.Score = matching.Score(places)
+	return
+}
+
 func FindBestPlanningSolutions(candidates []PlanningSolution, topSolutionsCount int64) []PlanningSolution {
 	if topSolutionsCount <= 0 {
 		topSolutionsCount = TopSolutionsCountDefault
@@ -88,7 +131,7 @@ func GenerateSolutions(context context.Context, matcher matching.Matcher, redisC
 	}
 
 	for mdIter.HasNext() {
-		curCandidate := CreateCandidate(mdIter, placeClusters)
+		curCandidate := CreatePlanningSolutionCandidate(mdIter, placeClusters)
 
 		solutions = append(solutions, curCandidate)
